tests/wpt/web-platform-tests/infrastructure/expected-fail/unhandled-rejection-following-subtest.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Unhandled rejection following subtest</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({ explicit_done: true });
window.addEventListener('unhandledrejection', () => { done(); }, {once: true});
test(function() {});
Promise.reject(new Error("error outside any setup or test"));
</script>
