LayoutTests/imported/w3c/web-platform-tests/infrastructure/expected-fail/unhandled-rejection-following-subtest-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: error outside any setup or test

PASS Unhandled rejection following subtest
